Save and Restore config back-up files #42
1. Copy static/config-backups/* to config-backups-rpmsave:
incoming package %pre scriptlet (first to run).
2. Copy config-backups-rpmsave/* to static/config-backups:
incoming package %posttrans (last to run and after build.sh).

It is important that we restore our config backup files after
build.sh as it depends on a non-existent static directory to
establish incoming static contents: see %post, in update mode,
which ensures the removal of the static directory.