Move build.sh execution from %posttrans to rockstor-build.service #61
Move the execution of build.sh from outside of the constraints of
the rpm packaging system, to the level of OS orchestration: systemd.

## Includes:
- Removal of redundant Poetry install/update. We now assume this of our
build host, and assert rockstor-build.service / build.sh as the sole
source of truth for this.
- Simplify %check scriptlet by similarly removing redundancy re GNUPG &
'pass' setup/config. Referencing rockstor-build.service / build.sh for
this build host assumption.