Update DB format from 13 to 17 re Django 5.2 compat #75
Djanog 5.2 has a minimum Postgres compatibility of 14.
Three jumps from our prior 13 to 16 was considered,
however there are known bugs in 16 re SSE, one of our
pending aims. Ergo 17 was chosen as it is also available
in our current OS targets of Leap 15.6 and TW/SR.

Adds another call to follow our existing `db_upgrade 10 13`,
but as `db_upgrade 13 17`, along with a change to the existing
update-alternatives to set a default of 17 for Postgres.